support old X86 CPU without AES-NI
Use slow implementation for old CPUs.
